organization/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12.0
django-mptt>=0.13.4
psycopg2-binary>=2.9.3
Faker>=13.3.4
gunicorn>=20.1.0
drf-yasg>=1.20.0
pytest>=7.0.0
pytest-django>=4.5.2
pytest-cov>=4.0.0
factory-boy>=3.2.1